Monitored process: Will now stop trying to terminate a process after the third attempt

+++ Core/general/monitoredProcess.cpp
@@ -260,8 +260,13 @@ namespace DeltaWorks
 				PostMessage(*w,WM_CLOSE,0,0);
 				//PostThreadMessageA(threadID,WM_CLOSE,0,0);
 			}
+			count_t numTries = 0;
 			if (WaitForSingleObject(infoOut.hProcess,2000) != 0)
+			{
+				if (++numTries > 3)
+					throw Except::Program::GeneralFault(CLOCATION,"Unable to terminate monitored process");
 				TerminateProcess(infoOut.hProcess,0);
+			}
 			Flush();
 		}
 	}
